collapse method.v1.0.ts into method.ts and create core/
