+++ src/core/entries.ts
