+++ src/core/resolution.ts
